+++ srcs/lexer_check_character_symbol.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   check_character_symbol.c                           :+:      :+:    :+:   */
+/*   lexer_check_character_symbol.c                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sejkim2 <sejkim2@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/13 18:07:22 by sejkim2           #+#    #+#             */
-/*   Updated: 2023/09/13 18:07:48 by sejkim2          ###   ########.fr       */
+/*   Updated: 2023/09/20 13:42:12 by sejkim2          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
