+++ srcs/lexer_error.c
@@ -1,43 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   main.c                                             :+:      :+:    :+:   */
+/*   lexer_error.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sejkim2 <sejkim2@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/08/25 15:52:52 by jaehyji           #+#    #+#             */
-/*   Updated: 2023/09/20 13:42:23 by sejkim2          ###   ########.fr       */
+/*   Created: 2023/09/20 12:47:00 by sejkim2           #+#    #+#             */
+/*   Updated: 2023/09/20 13:42:14 by sejkim2          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
-int	main(void)
+void lexer_error()
 {
-	char			*line;
-	t_linked_list	*list;
-
-	set_shell_signal();
-	while (1)
-	{
-		line = readline("minishell$ ");
-		if (line[0] == '\0')
-		{
-			free(line);
-			continue;
-		}
-		if (line)
-		{
-			add_history(line);
-			list = lexer(line);
-			parser(list); //check_syntax_errror
-			//execve()
-			// free_list(list);
-			free(line);
-			line = 0;
-		}
-		else
-			return (ctrl_d());
-	}
-	return (0);
+    printf("lexer error!\n");
+    exit(1);
 }
