+++ srcs/lexer_free_list.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   free_list.c                                        :+:      :+:    :+:   */
+/*   lexer_free_list.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sejkim2 <sejkim2@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 17:55:41 by sejkim2           #+#    #+#             */
-/*   Updated: 2023/09/06 19:17:16 by sejkim2          ###   ########.fr       */
+/*   Updated: 2023/09/20 13:42:16 by sejkim2          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
