+++ srcs/lexer_list.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   list.c                                             :+:      :+:    :+:   */
+/*   lexer_list.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sejkim2 <sejkim2@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 17:13:44 by sejkim2           #+#    #+#             */
-/*   Updated: 2023/09/14 20:08:50 by sejkim2          ###   ########.fr       */
+/*   Updated: 2023/09/20 13:42:19 by sejkim2          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
