+++ srcs/lexer_node.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   node.c                                             :+:      :+:    :+:   */
+/*   lexer_node.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sejkim2 <sejkim2@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/13 18:03:01 by sejkim2           #+#    #+#             */
-/*   Updated: 2023/09/20 12:14:54 by sejkim2          ###   ########.fr       */
+/*   Updated: 2023/09/20 13:42:20 by sejkim2          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
