+++ srcs/shell_signal.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   shell_signal.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jaehyji <jaehyji@student.42.fr>            +#+  +:+       +#+        */
+/*   By: sejkim2 <sejkim2@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 16:54:42 by jaehyji           #+#    #+#             */
-/*   Updated: 2023/09/06 20:17:47 by jaehyji          ###   ########.fr       */
+/*   Updated: 2023/09/20 13:42:27 by sejkim2          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
